requirements.txt
Django==1.8.2
djangorestframework==3.1.3
django-oauth-toolkit==0.8.2
django-phonenumber-field==0.7.2
Pillow==2.9.0
django-cors-headers==1.1.0
